relocate useAcceptedUploadFileType
